add enums to avoid magic strings
