+++ src/components/PeoplePage/PeoplePage.tsx
@@ -2,17 +2,13 @@ import { useEffect, useState } from 'react';
 import { Person } from '../../types';
 import { getPeople } from '../../api';
 import { Loader } from '../Loader/Loader';
-import { PeopleTable } from './PeopleTable';
-import { useParams } from 'react-router-dom';
 import { Errors } from '../../types/Errors';
+import { PeopleTable } from './PeopleTable';
 
 export const PeoplePage = () => {
   const [people, setPeople] = useState<Person[]>([]);
   const [isLoading, setIsLoading] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
-  const { personSlug } = useParams();
-
-  const currentPerson = people.find(person => person.slug === personSlug);
 
   useEffect(() => {
     setIsLoading(true);
@@ -41,37 +37,7 @@ export const PeoplePage = () => {
                   There are no people on the server
                 </p>
               )}
-              <table
-                data-cy="peopleTable"
-                className="table is-striped is-hoverable is-narrow is-fullwidth"
-              >
-                <thead>
-                  <tr>
-                    <th>Name</th>
-                    <th>Sex</th>
-                    <th>Born</th>
-                    <th>Died</th>
-                    <th>Mother</th>
-                    <th>Father</th>
-                  </tr>
-                </thead>
-
-                <tbody>
-                  {people.map(person => (
-                    <PeopleTable
-                      person={person}
-                      key={person.slug}
-                      currentPerson={currentPerson}
-                      mother={people.find(
-                        parent => parent.name === person.motherName,
-                      )}
-                      father={people.find(
-                        parent => parent.name === person.fatherName,
-                      )}
-                    />
-                  ))}
-                </tbody>
-              </table>
+              <PeopleTable people={people} />
             </>
           )}
         </div>
