--- src/components/PeoplePage/Table.tsx
@@ -1,12 +1,0 @@
-export const Table = () => (
-  <thead>
-    <tr>
-      <th>Name</th>
-      <th>Sex</th>
-      <th>Born</th>
-      <th>Died</th>
-      <th>Mother</th>
-      <th>Father</th>
-    </tr>
-  </thead>
-);
