+++ src/types/Sex.ts
@@ -1,0 +1,4 @@
+export enum Sex {
+  Male = 'm',
+  Female = 'f',
+}
